Day6/pyproject.toml
[project]
name = "day6"
version = "0.1.0"
description = ""
authors = [
    {name = "khonghao0109",email = "khonghao0109@gmail.com"}
]
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.135.3,<0.136.0)",
    "uvicorn (>=0.44.0,<0.45.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "bcrypt (<5)",
    "email-validator (>=2.3.0,<3.0.0)"
]

[tool.uv]
package = false


[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"
